updated checking for account and data

+++ cloud_server.py
@@ -14,47 +14,48 @@ server.listen(10)
 if not os.path.exists("accounts"):
     os.mkdir("accounts")
 
-x = 0
 while True:
     client,addr = server.accept()
-    x += 1
     print('Servicing client at %s'%addr[0])
     res = 'You have connected to %s, please stand by...'%host
     client.send(res.encode('UTF-8'))
 
     data_received = client.recv(1024)
-    
-    
-    success_mes = 'Successfully Received'
-    client.send(success_mes.encode('UTF-8'))
-
-
-    #try and catch for decode 
-
     data = json.loads(data_received)
-    
+
 
     directory = data["UID"]
     cwd = os.getcwd()
     newpath = os.path.join(cwd, "accounts", directory)
-    if not os.path.exists(newpath):
-       os.mkdir(newpath)
 
-    
-    working_file = os.path.join(newpath, data["File"])
-    f = open(working_file, 'w') 
-    f.write(data["Data"] + "\n")
-    f.close()
 
-    #check if there no data passed? 
+    if data["Data"] is None and not os.path.exists(newpath):
+      os.mkdir(newpath)
+      success_mes = 'Successfully created account, no data updated'
+      client.send(success_mes.encode('UTF-8'))  
     
-    #successfully updated message 
+    elif data["Data"] is None and os.path.exists(newpath):
+        success_mes = 'Welcome back, no data updated'
+        client.send(success_mes.encode('UTF-8')) 
+   
+    elif not os.path.exists(newpath):
+      os.mkdir(newpath)
+      working_file = os.path.join(newpath, data["File"])
+      f = open(working_file, 'w') 
+      f.write(data["Data"] + "\n")
+      f.close()
+      success_mes = 'Successfully created an account and saved the data'
+      client.send(success_mes.encode('UTF-8')) 
+
+    elif os.path.exists(newpath):
+      working_file = os.path.join(newpath, data["File"])
+      f = open(working_file, 'w') 
+      f.write(data["Data"] + "\n")
+      f.close()
+      success_mes = 'Welcome back, successfully updated the data'
+      client.send(success_mes.encode('UTF-8')) 
+    
 
     client.close()
-
-    #print (data_received)
-
-
-
 
 server.close()
